+++ adminalerts/tests/test_views.py
@@ -50,6 +50,7 @@ class AdminalertsViewTestBase(
         self.superuser = self.make_user('superuser')
         self.superuser.is_superuser = True
         self.superuser.is_staff = True
+        self.superuser.email = 'superuser@example.com'
         self.superuser.save()
         self.user_regular = self.make_user('user_regular')
         # No user
@@ -150,6 +151,7 @@ class TestAdminAlertCreateView(AdminalertsViewTestBase):
         self.assertIn(ALERT_MSG, mail.outbox[0].body)
         self.assertIn(EMAIL_DESC_LEGEND, mail.outbox[0].body)
         self.assertIn(ALERT_DESC, mail.outbox[0].body)
+        self.assertNotIn(self.superuser.email, mail.outbox[0].body)
 
     def test_post_no_description(self):
         """Test POST with no description"""
@@ -229,10 +231,27 @@ class TestAdminAlertCreateView(AdminalertsViewTestBase):
             mail.outbox[0].recipients(),
             [
                 settings.EMAIL_SENDER,
-                self.superuser.email,
-                self.user_regular.email,
                 ADD_EMAIL,
                 ADD_EMAIL2,
+                self.superuser.email,
+                self.user_regular.email,
+            ],
+        )
+
+    def test_post_no_primary_email(self):
+        """Test POST with no primary additional email on user"""
+        self.user_regular.email = ''
+        self.user_regular.save()
+        self.make_email(self.user_regular, ADD_EMAIL)
+        with self.login(self.superuser):
+            response = self.client.post(self.url, self._get_post_data())
+        self.assertEqual(response.status_code, 302)
+        self.assertEqual(
+            mail.outbox[0].recipients(),
+            [
+                settings.EMAIL_SENDER,
+                ADD_EMAIL,  # This should still be added
+                self.superuser.email,
             ],
         )
 
@@ -250,9 +269,9 @@ class TestAdminAlertCreateView(AdminalertsViewTestBase):
             mail.outbox[0].recipients(),
             [
                 settings.EMAIL_SENDER,
+                ADD_EMAIL,
                 self.superuser.email,
                 self.user_regular.email,
-                ADD_EMAIL,
             ],
         )
 
@@ -271,9 +290,9 @@ class TestAdminAlertCreateView(AdminalertsViewTestBase):
             mail.outbox[0].recipients(),
             [
                 settings.EMAIL_SENDER,
-                self.superuser.email,
                 'add1@example.com',
                 'add2@example.com',
+                self.superuser.email,
                 self.user_regular.email,
             ],
         )