+++ adminalerts/views.py
@@ -18,7 +18,7 @@ from django.views.generic.edit import ModelFormMixin
 
 # Projectroles dependency
 from projectroles.app_settings import AppSettingAPI
-from projectroles.email import get_email_user, get_user_addr, send_generic_mail
+from projectroles.email import get_user_addr, send_generic_mail
 from projectroles.views import (
     LoggedInPermissionMixin,
     HTTPRefererMixin,
@@ -93,15 +93,15 @@ class AdminAlertModifyMixin(ModelFormMixin):
     def _get_email_recipients(cls, alert):
         """Return list of email addresses for alert email recipients"""
         ret = []
-        for u in User.objects.order_by('email'):
+        for u in User.objects.all():
             if not app_settings.get(APP_NAME, 'notify_email_alert', user=u):
                 continue
-            if not u.email:
-                logger.warning('No email set for user: {}'.format(u.username))
-                continue
             user_emails = get_user_addr(u)
+            if not user_emails:
+                logger.warning('No emails set for user: {}'.format(u.username))
+                continue
             ret += [e for e in user_emails if e not in ret]
-        return ret
+        return sorted(ret)
 
     def _send_email(self, alert, action):
         """
@@ -116,7 +116,7 @@ class AdminAlertModifyMixin(ModelFormMixin):
         )
         body = EMAIL_BODY.format(
             action=action,
-            issuer=get_email_user(alert.user),
+            issuer=alert.user.get_display_name(inc_user=True),
             message=alert.message,
         )
         if alert.description: