+++ docs/source/conf.py
@@ -29,7 +29,7 @@ author = 'BIH Core Unit Bioinformatics'
 # The short X.Y version
 version = '1.1'
 # The full version, including alpha/beta/rc tags
-release = '1.1.1'
+release = '1.1.2'
 
 
 # -- General configuration ---------------------------------------------------